@@ -8,11 +8,11 @@ buildscript {
 
         libVersionCode = libPatch + (libMinor * 100) + (libMajor * 10000)
 
-        kotlin_version = '1.3.72'
-        core_ktx_version = '1.3.0'
-        exoplayer_version = '2.11.7'
-        appcompat_version = '1.1.0'
-        constraintlayout_version = '1.1.3'
+        kotlin_version = '1.4.20'
+        core_ktx_version = '1.3.2'
+        exoplayer_version = '2.12.1'
+        appcompat_version = '1.2.0'
+        constraintlayout_version = '2.0.4'
     }
 
     repositories {
@@ -21,7 +21,7 @@ buildscript {
 
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.0.1'
+        classpath 'com.android.tools.build:gradle:4.1.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
