@@ -5,13 +5,11 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.view.View
 import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import com.github.vkay94.dtpv.DoubleTapPlayerView
-import com.google.android.exoplayer2.DefaultLoadControl
-import com.google.android.exoplayer2.LoadControl
-import com.google.android.exoplayer2.Player
-import com.google.android.exoplayer2.SimpleExoPlayer
+import com.google.android.exoplayer2.*
 import com.google.android.exoplayer2.source.ProgressiveMediaSource
 import com.google.android.exoplayer2.upstream.DefaultBandwidthMeter
 import com.google.android.exoplayer2.upstream.DefaultDataSourceFactory
@@ -36,10 +34,12 @@ open class BaseVideoActivity : AppCompatActivity() {
             DefaultBandwidthMeter.Builder(this@BaseVideoActivity).build()
         )
         val videoSource = ProgressiveMediaSource.Factory(dataSourceFactory, Mp4ExtractorFactory())
-            .createMediaSource(mUri)
+            .createMediaSource(MediaItem.fromUri(mUri))
 
-        player?.prepare(videoSource)
-        player?.playWhenReady = true
+        player?.apply {
+            setMediaSource(videoSource)
+            prepare()
+        }
     }
 
     fun initializePlayer() {
@@ -51,7 +51,7 @@ open class BaseVideoActivity : AppCompatActivity() {
                     MIN_PLAYBACK_START_BUFFER,
                     MIN_PLAYBACK_RESUME_BUFFER
                 )
-                .createDefaultLoadControl()
+                .build()
 
             player = SimpleExoPlayer.Builder(this)
                 .setLoadControl(loadControl)
@@ -69,48 +69,39 @@ open class BaseVideoActivity : AppCompatActivity() {
         }
     }
 
-    fun pausePlayer() {
-        if (player != null) {
-            player?.playWhenReady = false
-            player?.playbackState
-        }
-    }
-
-    fun resumePlayer() {
-        if (player != null) {
-            player?.playWhenReady = true
-            player?.playbackState
-        }
-    }
-
     override fun onPause() {
         super.onPause()
-        pausePlayer()
+        player?.pause()
     }
 
     override fun onRestart() {
         super.onRestart()
         if (player?.playbackState == Player.STATE_READY && player?.playWhenReady!!)
-            resumePlayer()
+            player?.play()
     }
 
     override fun onDestroy() {
         super.onDestroy()
-
-        if (player != null) {
-            player?.release()
-            player = null
-        }
+        releasePlayer()
     }
 
     fun setFullscreen(fullscreen: Boolean) {
         if (fullscreen) {
-            this.window.setFlags(
+            window.setFlags(
                 WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN
             )
+            window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE
         } else {
-            this.window.clearFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN)
+            window.clearFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN)
+            if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.KITKAT) {
+                window.decorView.systemUiVisibility = (View.SYSTEM_UI_FLAG_FULLSCREEN
+                        and View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                        and View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)
+            } else {
+                window.decorView.systemUiVisibility = (View.SYSTEM_UI_FLAG_FULLSCREEN
+                        and View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)
+            }
         }
     }
 
