@@ -8,29 +8,45 @@ import android.view.Menu
 import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
-import androidx.lifecycle.Observer
+import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.lifecycle.ViewModelProvider
+import com.github.vkay94.doubletapplayerviewexample.databinding.ActivityMainBinding
+import com.github.vkay94.doubletapplayerviewexample.databinding.ExoPlaybackControlViewYtBinding
 import com.github.vkay94.doubletapplayerviewexample.fragments.PageViewModel
 import com.github.vkay94.doubletapplayerviewexample.fragments.SecondsViewFragment
 import com.github.vkay94.doubletapplayerviewexample.fragments.SectionsPagerAdapter
 import com.github.vkay94.doubletapplayerviewexample.fragments.ShapeFragment
 import com.github.vkay94.dtpv.youtube.YouTubeOverlay
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.exo_playback_control_view_yt.*
 
 class MainActivity : BaseVideoActivity() {
 
     private var isVideoFullscreen = false
     private var currentVideoId = -1
 
+    // View bindings
+    private lateinit var binding: ActivityMainBinding
+    private lateinit var controlsBinding: ExoPlaybackControlViewYtBinding
+
     private lateinit var viewModel: PageViewModel
+
+    // Views (to eliminate repeating 'binding.'-prefixes)
+    private lateinit var ytOverlay: YouTubeOverlay
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        setSupportActionBar(toolbar)
 
-        this.videoPlayer = previewPlayerView
+        // Setup layout views with View binding
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        val view = binding.root
+        val controls = view.findViewById<ConstraintLayout>(R.id.exo_controls_root)
+        controlsBinding = ExoPlaybackControlViewYtBinding.bind(controls)
+        setContentView(view)
+
+        ytOverlay = binding.ytOverlay
+        videoPlayer = binding.previewPlayerView
+
+        setSupportActionBar(binding.toolbar)
+
         initDoubleTapPlayerView()
         initViewModel()
         startNextVideo()
@@ -39,10 +55,10 @@ class MainActivity : BaseVideoActivity() {
         sectionsPagerAdapter.addFragment(ShapeFragment.newInstance())
         sectionsPagerAdapter.addFragment(SecondsViewFragment.newInstance())
 
-        view_pager.adapter = sectionsPagerAdapter
-        tabs.setupWithViewPager(view_pager)
+        binding.viewPager.adapter = sectionsPagerAdapter
+        binding.tabs.setupWithViewPager(binding.viewPager)
 
-        fullscreen_button.setOnClickListener {
+        controlsBinding.fullscreenButton.setOnClickListener {
             toggleFullscreen()
         }
     }
@@ -53,16 +69,16 @@ class MainActivity : BaseVideoActivity() {
             //.playerView(previewPlayerView)
             .performListener(object : YouTubeOverlay.PerformListener {
                 override fun onAnimationStart() {
-                    previewPlayerView.useController = false
+                    binding.previewPlayerView.useController = false
                     ytOverlay.visibility = View.VISIBLE
                 }
                 override fun onAnimationEnd() {
                     ytOverlay.visibility = View.GONE
-                    previewPlayerView.useController = true
+                    binding.previewPlayerView.useController = true
                 }
             })
 
-        previewPlayerView.doubleTapDelay = 800
+        binding.previewPlayerView.doubleTapDelay = 800
         // Uncomment this line if the PlayerDoubleTapListener is not set via XML
         // previewPlayerView.controller(ytOverlay)
     }
@@ -78,28 +94,28 @@ class MainActivity : BaseVideoActivity() {
             iconSpeed.value = ytOverlay.iconAnimationDuration
         }
 
-        viewModel.circleExpandDuration.observe(this, Observer {
+        viewModel.circleExpandDuration.observe(this, {
             ytOverlay.animationDuration(it)
         })
-        viewModel.arcSize.observe(this, Observer {
+        viewModel.arcSize.observe(this, {
             ytOverlay.arcSize(DataAndUtils.dpToPx(this@MainActivity, it.toFloat()))
         })
-        viewModel.fontSize.observe(this, Observer {
+        viewModel.fontSize.observe(this, {
             TextViewStyler().textSize(it).applyTo(ytOverlay.secondsTextView)
         })
-        viewModel.typeFace.observe(this, Observer {
+        viewModel.typeFace.observe(this, {
             TextViewStyler().textStyle(it).applyTo(ytOverlay.secondsTextView)
         })
-        viewModel.tapCircleColor.observe(this, Observer {
+        viewModel.tapCircleColor.observe(this, {
             ytOverlay.tapCircleColorInt(it)
         })
-        viewModel.circleBackgroundColor.observe(this, Observer {
+        viewModel.circleBackgroundColor.observe(this, {
             ytOverlay.circleBackgroundColorInt(it)
         })
-        viewModel.secondsIcon.observe(this, Observer {
+        viewModel.secondsIcon.observe(this, {
             ytOverlay.icon(it)
         })
-        viewModel.iconSpeed.observe(this, Observer {
+        viewModel.iconSpeed.observe(this, {
             ytOverlay.iconAnimationDuration(it)
         })
     }
@@ -111,27 +127,23 @@ class MainActivity : BaseVideoActivity() {
 
         currentVideoId = (currentVideoId + 1).rem(DataAndUtils.videoList.size)
         buildMediaSource(Uri.parse(DataAndUtils.videoList[currentVideoId]))
+        player?.play()
     }
 
     private fun toggleFullscreen() {
         if (isVideoFullscreen) {
             setFullscreen(false)
-            window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_VISIBLE;
-            if(supportActionBar != null){
-                supportActionBar?.show();
+            if (supportActionBar != null) {
+                supportActionBar?.show()
             }
-            requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT;
+            requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
             isVideoFullscreen = false
         } else {
             setFullscreen(true)
-            window.decorView.systemUiVisibility = (View.SYSTEM_UI_FLAG_FULLSCREEN
-                    and View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                    and View.SYSTEM_UI_FLAG_HIDE_NAVIGATION)
-
-            if(supportActionBar != null){
-                supportActionBar?.hide();
+            if (supportActionBar != null) {
+                supportActionBar?.hide()
             }
-            requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE;
+            requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
             isVideoFullscreen = true
         }
     }
